backend/pyproject.toml
[project]
name = "giterdone-backend"
version = "0.1.0"
description = "Giterdone Todo App Backend - Django REST API"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2,<5.3",
    "django-cors-headers>=4.6,<5.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.4,<6.0",
    "psycopg[binary,pool]>=3.2,<4.0",
    "pyotp>=2.9,<3.0",
    "python-dotenv>=1.0,<2.0",
    "qrcode[pil]>=8.0,<9.0",
    "webauthn>=2.3,<3.0",
    # Production dependencies
    "gunicorn>=23.0,<24.0",
    "whitenoise>=6.8,<7.0",
    "django-ratelimit>=4.1,<5.0",
    "redis>=5.2,<6.0",
    "pytest-xdist>=3.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "giterdone.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["users/tests", "todos/tests"]
addopts = ["--tb=short", "--strict-markers", "--cov=.", "--cov-report=html", "--cov-report=term-missing"]

[tool.coverage.run]
source = ["."]
omit = ["*/migrations/*", "*/tests/*", "manage.py", "*/venv/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "coverage>=7.6,<8.0",
    "pytest>=8.3,<9.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.9,<5.0",
]
